[marlowe-integration-tests] renamed modules
