+++ marlowe-integration-tests/test/Language/Marlowe/Runtime/Web/Contracts/Transactions/Transaction/Put.hs
@@ -1,23 +1,27 @@
-module Language.Marlowe.Runtime.Web.Contracts.Contract.PutContract where
+module Language.Marlowe.Runtime.Web.Contracts.Transactions.Transaction.Put where
 
 import Control.Monad.IO.Class (MonadIO(liftIO))
 import qualified Data.Map.Strict as Map
 import qualified Data.Set as Set
 import Data.Time (getCurrentTime, secondsToNominalDiffTime)
+import Language.Marlowe.Core.V1.Semantics.Types (Input(NormalInput), InputContent(IDeposit))
+import Language.Marlowe.Extended.V1 (ada)
 import Language.Marlowe.Runtime.Integration.Common
+  (Wallet(Wallet, addresses, signingKeys), expectJust, getGenesisWallet, runIntegrationTest, runWebClient)
 import Language.Marlowe.Runtime.Integration.StandardContract (standardContract)
 import Language.Marlowe.Runtime.Plutus.V2.Api (toPlutusAddress)
 import Language.Marlowe.Runtime.Transaction.Api (WalletAddresses(..))
 import Language.Marlowe.Runtime.Web (RoleTokenConfig(RoleTokenSimple))
 import qualified Language.Marlowe.Runtime.Web as Web
-import Language.Marlowe.Runtime.Web.Client (postContract, putContract)
-import Language.Marlowe.Runtime.Web.Common (signShelleyTransaction')
+import Language.Marlowe.Runtime.Web.Client (postContract, postTransaction, putTransaction)
+import Language.Marlowe.Runtime.Web.Common (signShelleyTransaction', submitContract)
 import Language.Marlowe.Runtime.Web.Server.DTO (ToDTO(toDTO))
 import Test.Hspec (Spec, describe, it)
 import Test.Integration.Marlowe.Local (withLocalMarloweRuntime)
 
+
 spec :: Spec
-spec = describe "POST /contracts/{contractId}/transactions" do
+spec = describe "PUT /contracts/{contractId}/transactions/{transaction}" do
   it "returns the transaction header" $ withLocalMarloweRuntime $ runIntegrationTest do
     partyAWallet@Wallet{signingKeys} <- getGenesisWallet 0
     partyBWallet <- getGenesisWallet 1
@@ -33,22 +37,41 @@ spec = describe "POST /contracts/{contractId}/transactions" do
       partyBAddress <- liftIO $ expectJust "Failed to convert party B address" $ toPlutusAddress $ changeAddress partyBWalletAddresses
       now <- liftIO getCurrentTime
 
-      let (contract, _, _) = standardContract partyBAddress now $ secondsToNominalDiffTime 100
+      let (contract, partyA, _) = standardContract partyBAddress now $ secondsToNominalDiffTime 100
 
-      Web.CreateTxEnvelope{contractId, txEnvelope} <- postContract
+      contractCreated@Web.CreateTxEnvelope{contractId} <- postContract
         partyAWebChangeAddress
         (Just partyAWebExtraAddresses)
         (Just partyAWebCollataralUtxos)
         Web.PostContractsRequest
           { metadata = mempty
           , version = Web.V1
-          , roles = Just $ Web.Mint $ Map.singleton "PartyA" $ RoleTokenSimple partyAWebChangeAddress
+          , roles = Just $ Web.Mint $ Map.singleton "Party A" $ RoleTokenSimple partyAWebChangeAddress
           , contract = contract
           , minUTxODeposit = 2_000_000
           , tags = mempty
           }
-      signedCreateTx <- liftIO $ signShelleyTransaction' txEnvelope signingKeys
-      putContract contractId signedCreateTx
+
+      _ <- submitContract partyAWallet contractCreated
+
+
+      let inputs = [NormalInput $ IDeposit partyA partyA ada 100_000_000]
+
+      Web.ApplyInputsTxEnvelope{transactionId, txEnvelope} <- postTransaction
+        partyAWebChangeAddress
+        (Just partyAWebExtraAddresses)
+        (Just partyAWebCollataralUtxos)
+        contractId
+        Web.PostTransactionsRequest
+          { version = Web.V1
+          , metadata = mempty
+          , invalidBefore = Nothing
+          , invalidHereafter = Nothing
+          , inputs
+          , tags = mempty
+          }
+      applyTx <- liftIO $ signShelleyTransaction' txEnvelope signingKeys
+      putTransaction contractId transactionId applyTx
     case result of
       Left _ ->  fail $ "Expected 200 response code - got " <> show result
       Right () ->  pure ()
