+++ marlowe-integration-tests/test/Language/Marlowe/Runtime/WebSpec.hs
@@ -2,21 +2,22 @@
 module Language.Marlowe.Runtime.WebSpec where
 
 
-import qualified Language.Marlowe.Runtime.Web.Contracts.GetContracts as Contracts.Get
-
-import qualified Language.Marlowe.Runtime.Web.Contracts.Contract.GetContract as Contracts.Contract.Get
-import qualified Language.Marlowe.Runtime.Web.Contracts.Contract.PostContract as Contracts.Contract.Post
-import qualified Language.Marlowe.Runtime.Web.Contracts.Contract.PutContract as Contracts.Contract.Put
-
-import qualified Language.Marlowe.Runtime.Web.Contracts.Transactions.GetTransactions as Contracts.Transactions.Get
-
-import qualified Language.Marlowe.Runtime.Web.Contracts.Transactions.Transaction.GetTransaction as Contracts.Transactions.Transaction.Get
-import qualified Language.Marlowe.Runtime.Web.Contracts.Transactions.Transaction.PostTransaction as Contracts.Transactions.Transaction.Post
-import qualified Language.Marlowe.Runtime.Web.Contracts.Transactions.Transaction.PutTransaction as Contracts.Transactions.Transaction.Put
+import qualified Language.Marlowe.Runtime.Web.Contracts.Get as Contracts.Get
 
 
-import qualified Language.Marlowe.Runtime.Web.Withdrawal.PostWithdrawal as Withdrawal.Post
-import qualified Language.Marlowe.Runtime.Web.Withdrawal.PutWithdrawal as Withdrawal.Put
+import qualified Language.Marlowe.Runtime.Web.Contracts.Contract.Get as Contracts.Contract.Get
+import qualified Language.Marlowe.Runtime.Web.Contracts.Contract.Post as Contracts.Contract.Post
+import qualified Language.Marlowe.Runtime.Web.Contracts.Contract.Put as Contracts.Contract.Put
+
+import qualified Language.Marlowe.Runtime.Web.Contracts.Transactions.Get as Contracts.Transactions.Get
+
+import qualified Language.Marlowe.Runtime.Web.Contracts.Transactions.Transaction.Get as Contracts.Transactions.Transaction.Get
+import qualified Language.Marlowe.Runtime.Web.Contracts.Transactions.Transaction.Post as Contracts.Transactions.Transaction.Post
+import qualified Language.Marlowe.Runtime.Web.Contracts.Transactions.Transaction.Put as Contracts.Transactions.Transaction.Put
+
+
+import qualified Language.Marlowe.Runtime.Web.Withdrawal.Post as Withdrawal.Post
+import qualified Language.Marlowe.Runtime.Web.Withdrawal.Put as Withdrawal.Put
 
 
 import Test.Hspec (Spec, describe)
